# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  trips.ChecklistCategory:
    enum:
    - booking
    - documents
    - packing
    - money
    - health
    - other
    type: string
    x-enum-varnames:
    - ChecklistCategoryBooking
    - ChecklistCategoryDocuments
    - ChecklistCategoryPacking
    - ChecklistCategoryMoney
    - ChecklistCategoryHealth
    - ChecklistCategoryOther
  trips.ChecklistItem:
    properties:
      assignee_id:
        type: string
      category:
        allOf:
        - $ref: '#/definitions/trips.ChecklistCategory'
        example: documents
      completed_at:
        type: string
      created_at:
        type: string
      due_date:
        type: string
      hop_id:
        type: string
      id:
        type: string
      is_completed:
        type: boolean
      notes:
        type: string
      sort_order:
        type: integer
      title:
        example: Apply for French visa
        type: string
      trip_plan:
        type: string
      updated_at:
        type: string
    type: object
  trips.ChecklistItemRequest:
    properties:
      assignee_id:
        type: string
      category:
        allOf:
        - $ref: '#/definitions/trips.ChecklistCategory'
        example: documents
      due_date:
        type: string
      hop_id:
        type: string
      is_completed:
        type: boolean
      notes:
        type: string
      sort_order:
        type: integer
      title:
        example: Apply for French visa
        type: string
    required:
    - title
    type: object
  trips.ContactRole:
    enum:
    - hotel
    - guide
    - driver
    - embassy
    - insurance
    - emergency
    - other
    type: string
    x-enum-varnames:
    - ContactRoleHotel
    - ContactRoleGuide
    - ContactRoleDriver
    - ContactRoleEmbassy
    - ContactRoleInsurance
    - ContactRoleEmergency
    - ContactRoleOther
...
  trips.TripContact:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        example: reservations@tajhotels.com
        type: string
      hop_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        example: Taj Hotel Front Desk
        type: string
      notes:
        type: string
      phone:
        example: +91-22-66651234
        type: string
      role:
        allOf:
        - $ref: '#/definitions/trips.ContactRole'
        example: hotel
      trip_plan:
        type: string
      updated_at:
        type: string
    type: object
  trips.TripContactRequest:
    properties:
      address:
        type: string
      email:
        type: string
      hop_id:
        type: string
      is_active:
        type: boolean
      name:
        example: Taj Hotel Front Desk
        type: string
      notes:
        type: string
      phone:
        example: +91-22-66651234
        type: string
      role:
        allOf:
        - $ref: '#/definitions/trips.ContactRole'
        example: hotel
    required:
    - name
    - role
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for managing trip plans, hops, and stays
  title: Trip Planner API
  version: "1.0"
paths:
...
  /checklist/{id}:
    delete:
      parameters:
      - description: Checklist Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Delete a checklist item
      tags:
      - checklist
    put:
      consumes:
      - application/json
      parameters:
      - description: Checklist Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated item
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/trips.ChecklistItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trips.ChecklistItem'
      security:
      - BearerAuth: []
      summary: Update a checklist item
      tags:
      - checklist
  /checklist/{id}/toggle:
    patch:
      description: Flips is_completed; sets completed_at when marking done, clears
        when undoing.
      parameters:
      - description: Checklist Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trips.ChecklistItem'
      security:
      - BearerAuth: []
      summary: Toggle completion of a checklist item
      tags:
      - checklist
  /contacts/{id}:
    delete:
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerAuth: []
      summary: Soft-delete a trip contact
      tags:
      - contacts
    put:
      consumes:
      - application/json
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated contact
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/trips.TripContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trips.TripContact'
      security:
      - BearerAuth: []
      summary: Update a trip contact
      tags:
      - contacts
...
  /trip/{id}/checklist:
    get:
      description: Returns all checklist items, optionally filtered by category or
        hop.
      parameters:
      - description: Trip Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by completion status
        in: query
        name: completed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: items
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List checklist items for a trip
      tags:
      - checklist
    post:
      consumes:
      - application/json
      parameters:
      - description: Trip Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Checklist item
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/trips.ChecklistItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/trips.ChecklistItem'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a checklist item to a trip
      tags:
      - checklist
  /trip/{id}/contacts:
    get:
      description: Returns all active contacts for a trip plan.
      parameters:
      - description: Trip Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by role (hotel, guide, driver, embassy, insurance, emergency,
          other)
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: contacts
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List contacts for a trip
      tags:
      - contacts
    post:
      consumes:
      - application/json
      parameters:
      - description: Trip Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Contact details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/trips.TripContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/trips.TripContact'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a contact to a trip
      tags:
      - contacts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"